Move classes out of tests/utils/__init__.py to their own modules
